services/shared/requirements.txt
# Shared dependencies for all microservices
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
python-dotenv==1.0.0
requests==2.31.0
PyJWT==2.8.0
